+++ lionengine-core/src/main/java/com/b3dgs/lionengine/geom/RectangleImpl.java
@@ -77,7 +77,7 @@ final class RectangleImpl
     }
 
     @Override
-    public boolean contains(int x, int y)
+    public boolean contains(double x, double y)
     {
         return x >= this.x && y >= this.y && x < this.x + width && y < this.y + height;
     }
